Promoted "Connected" to the main app frame
Instead of having two completely separate screens for the keyboard selector and
everything else, bring all of them under the same frame.

This way we have a uniform look, and can provide both the "Settings" and the
"Exit from Chrysalis" menus on every screen.

Fixes #125.
